+++ core/admin.py
@@ -1,6 +1,6 @@
 from django.contrib import admin
 
-from .models import SaleEntry, Balance, Gift, Cost, HipShipper, ReturnedSale
+from .models import SaleEntry, Balance, Gift, Cost, HipShipper, ReturnedSale, Preferences
 
 admin.site.register(SaleEntry)
 admin.site.register(Balance)
@@ -8,3 +8,4 @@ admin.site.register(Gift)
 admin.site.register(Cost)
 admin.site.register(HipShipper)
 admin.site.register(ReturnedSale)
+admin.site.register(Preferences)