+++ static/js/panel.js
@@ -195,7 +195,6 @@ function filter_gifts_by_date(){
                         day = date.split("-")[2]
                         
                         tr_style = "<tr>";
-                        console.log(item.fields.is_gift);
                         if (!item.fields.is_gift){
                             tr_style = "<tr id='bootstrap-overrides' class='warningrow'>"
                         }
@@ -691,6 +690,54 @@ $(document).ready(function(){
     sum_sales_table();
 
     calc_total_date_profit();
+
+    $(".lock-class").on("click", function(e){
+        e.preventDefault();
+
+        console.log($(this).hasClass("bg-success"));
+
+        $.ajax({
+            url: "/toggle_paypal_editable",
+            type: "POST",
+            data: {
+                value: $(this).hasClass("bg-success")
+            },
+            success: function(response){
+                $(".lock-class").toggleClass("bg-success bg-danger");
+
+                if ($(".lock-class").hasClass("bg-success")){
+                    // unlocked
+                    $(".lock-class").html("<i class='fa fa-lock-open fa-lg'></i>")
+                    paypal_balance_value = $("#div_paypal_balance").html()
+                    $("#div_paypal_balance").html(`
+                     <input type='number' id='txt_paypal_balance' class='form-control editable-paypal-balance'
+                     value='` + paypal_balance_value + `' data-lastvalue='` + paypal_balance_value + `' />`)
+                }
+                else{
+                    // locked
+                    $(".lock-class").html("<i class='fa fa-lock fa-lg'></i>")
+                    lastvalue = $("#txt_paypal_balance").data("lastvalue")
+                    curvalue = $("#txt_paypal_balance").val()
+                    
+                    if (lastvalue != curvalue){
+                        console.log(curvalue)
+                        $.ajax({
+                            url: '/update_paypal_balance',
+                            type: 'POST',
+                            data: {
+                                value: curvalue
+                            },
+                            success: function(response){
+                                $("#div_paypal_balance").html(curvalue);
+                            }
+                        })
+                    }
+                    
+                }
+            }
+        })
+        
+    });
 
     // sales count
     $("#sales-count").html("Sales: " + parseFloat($("#table_sales > tbody > tr").length - 1));